Split off lp_uart to submodule
